Wire node split into csched and let it run up to commit, but fail before commit. Fix busycnt accounting for compaction error path. Remove cn_node_size_{lo,hi} rparams in favor of cn_split_size rparam, Reduce size of unnecessarily large kvs rparams . Trigger per-node gc when a node reaches 15x garbage. Remove cn_tstate remnants. Move tn_rspills fields from cn_tree_node to cn_tree. Unify RBT_, jtype_, wtype_, in favor of wtype_. Add macros to specify default csched params. Add additional cn comp rules and unify naming with wtype_.

+++ lib/cn/kvset_internal.h
@@ -75,8 +75,8 @@ struct kvset {
     struct mpool *ks_mp;
     u32           ks_pfx_len; /* cn tree pfx_len */
     u32           ks_sfx_len; /* cn tree sfx_len */
-    u16           ks_node_level;
-    u16           ks_vminlvl;
+    uint8_t       ks_node_level;
+    uint8_t       ks_vminlvl;
     uint16_t      ks_comp_rule;
     u64           ks_nodeid;
     u32           ks_vmin;
